+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Pageinator (PagingControl)
 
 ## [Unreleased]
+
+## [1.8.0] - 2021-08-05
 ### Added
 * StackPanel, a control which behaves like the original stack panel but with the additional property for a spacing between the children
 * UniformPanel, A UniformGrid with only one row or one column, depending on the orientation, which adds a spacing between the items